+++ CHANGES.md
@@ -1,2 +1,14 @@
 # Changes
 
+## v0.1.0-a2
+
+-   Adapt APCu driver to PHP 7.1.x (Optional arguments type literal).
+-   Fixed the `getMulti` method in APCu driver.
+-   New methods:
+
+    - `getMultiString`
+    - `getMultiInt`
+    - `getMultiFloat`
+    - `setMultiString`
+    - `setMultiInt`
+    - `setMultiFloat`